+++ pg-core/src/java/org/pg/util/SQLTool.java
@@ -1,7 +1,9 @@
 package org.pg.util;
 
+import clojure.lang.Named;
 import org.pg.Const;
 import org.pg.enums.TxLevel;
+import org.pg.error.PGError;
 
 public final class SQLTool {
 
@@ -15,26 +17,28 @@ public final class SQLTool {
         return "SET TRANSACTION ISOLATION LEVEL " + level.getCode();
     }
 
-    public static String fullTypeName(final String type) {
-        String schema;
-        String name;
-        final String[] parts = type.split("\\.", 2);
-        if (parts.length == 1) {
-            schema = Const.defaultSchema;
-            name = parts[0];
-        } else {
-            schema = parts[0];
-            name = parts[1];
-        }
-        return schema + '.' + name;
+    public static String fullTypeName(final Object type) {
+        final String[] parts = splitType(type);
+        return parts[0] + '.' + parts[1];
     }
 
-    public static String[] splitType(final String type) {
-        final String[] parts = type.split("\\.", 2);
-        if (parts.length == 1) {
-            return new String[]{Const.defaultSchema, parts[0]};
+    public static String[] splitType(final Object type) {
+        if (type instanceof String s) {
+            final String[] parts = s.split("\\.", 2);
+            if (parts.length == 1) {
+                return new String[]{Const.defaultSchema, parts[0]};
+            } else {
+                return parts;
+            }
+        } else if (type instanceof Named nm) {
+            String schema;
+            schema = nm.getNamespace();
+            if (schema == null) {
+                schema = Const.defaultSchema;
+            }
+            return new String[]{schema, nm.getName()};
         } else {
-            return parts;
+            throw new PGError("unsupported postgres type, %s", TypeTool.repr(type));
         }
     }